KEYBOARD CONNECTED!!
Connected the keyboard to the game. When the user type a guess the game
checks against the puzzle and if its correct it displays the word. Need
to create the user interface and then write incorrect guess code.

TODO
2 Player Interface
Play Screen
proper name displaying and creating in player screens
category choosing